pyproject.toml
[project]
name = "weather-apis"
version = "0.1.0"
description = "Backend-only Django/DRF service for Weather APIs (NASA, Meteo, OpenWeather)."
requires-python = ">=3.11"
dependencies = [
  "Django>=5.1,<5.2",
  "dj-database-url>=2.3,<2.4",
  "djangorestframework>=3.15,<3.16",
  "djangorestframework-simplejwt>=5.3,<5.4",
  "drf-spectacular>=0.27,<0.28",
  "django-prometheus>=2.3,<2.4",
  "django-redis>=5.4,<5.5",
  "httpx>=0.27,<0.28",
  "python-dotenv>=1.0,<2.0",
  "django-environ>=0.11,<0.12",
]

[tool.black]
line-length = 79
target-version = ["py311"]
extend-exclude = '''
/(
  \.venv
| venv
| node_modules
| staticfiles
| dist
| build
| \.git
)/
'''

[tool.ruff]
line-length = 79
target-version = "py311"
exclude = ["**/migrations/**"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "S", "ANN"]
ignore = ["ANN401","ANN102","ANN101"]

[tool.bandit]
skips = ["B101"]
exclude_dirs = [
  ".venv",
  "venv",
  "migrations",
  "node_modules",
  ".git",
]
severity = "MEDIUM"
confidence = "HIGH"

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
disallow_any_unimported = true
ignore_missing_imports = false
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
explicit_package_bases = true
exclude = "(.venv|venv|node_modules)"

[tool.django-stubs]
django_settings_module = "config.mypy_settings"

[[tool.mypy.overrides]]
module = ["*.migrations.*"]
ignore_errors = true

[[tool.mypy.overrides]]
module = ["environ"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["celery", "celery.schedules"]
ignore_missing_imports = true
